% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_chromosome.R
\name{binary_chromosome_to_numeric}
\alias{binary_chromosome_to_numeric}
\title{Convert binary chromosome to numeric value}
\usage{
binary_chromosome_to_numeric(chromosome, min = 0, max = 1)
}
\arguments{
\item{chromosome}{Binary chromosome}

\item{min}{Minimal value which chromosome can represent}

\item{max}{Maximal value which chromosome can represent}
}
\value{
Numeric value within the range specified by \code{min}, \code{max}
}
\description{
Convert binary chromosome to numeric value.
}
\examples{
# Generate a random 24 bit chromosome
chromosome <- sample(c(TRUE, FALSE), 24, TRUE)
Convert the chromosome to a numeric value within the range [-10, 10]
binary_chromosome_to_numeric(chromosome, -10, 10)

}
