% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_objective_benchmarks.R
\name{easom_function}
\alias{easom_function}
\title{Easom benchmark function}
\usage{
easom_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Function value
}
\description{
Easom benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
