% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_objective_benchmarks.R
\name{fonseca_fleming_function}
\alias{fonseca_fleming_function}
\title{Fonseca-Fleming benchmark function}
\usage{
fonseca_fleming_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Vector with two objective functions values
}
\description{
Fonseca-Fleming benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
