% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_objective_benchmarks.R
\name{himmelblau_function}
\alias{himmelblau_function}
\title{Himmelblau benchmark function}
\usage{
himmelblau_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Function value

 @export
}
\description{
Himmelblau benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
