% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-dominance.R
\name{is_nondominated}
\alias{is_nondominated}
\title{Check if solution in nondominated}
\usage{
is_nondominated(x, solutions, minimize = TRUE)
}
\arguments{
\item{x}{Numeric vector with objective functions values for solution \code{x}}

\item{solutions}{List or matrix with objective functions values for \code{solutions}.
In matrix each row contains values for one solution}

\item{minimize}{If objective functions should be minimized}
}
\value{
Logical value
}
\description{
Check if solution \code{x} is not dominated by any other solution in the set \code{solutions}
}
\examples{
x <- c(1, 2, 2)
solutions <- list(c(1, 5, 5), c(8, 4, 8), c(3, 2, 2))
is_nondominated(x, solutions)

}
