% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{kursawe_function}
\alias{kursawe_function}
\title{Kursawe benchmark function}
\usage{
kursawe_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Vector with two objective functions values
}
\description{
Kursawe benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
