% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga2.R
\name{nsga2}
\alias{nsga2}
\title{NSGAII algorithm}
\usage{
nsga2(objective_functions_list, chromosome_size,
  chromosome_type = "binary",
  population_size = length(objective_functions_list) * 40,
  number_of_iterations = 100, nc = 2, mutation_probability = 0.05,
  uniform_mutation_sd = 0.01)
}
\arguments{
\item{objective_functions_list}{List of objective functions}

\item{chromosome_size}{Size of chromosome which represents candidate solutions}

\item{chromosome_type}{Chromosome type ("binary" or "numeric")}

\item{population_size}{Number of solutions evaluated in one iteration of genetic algorithm}

\item{number_of_iterations}{Number of iterations (generations) of genetic algorithm}

\item{nc}{NC for SBX crossover (valid if "numeric" chromosome is used)}

\item{mutation_probability}{Probability of mutation (valid if "binary" chromosome is used)}

\item{uniform_mutation_sd}{Standard deviation of mutation (valid if "numeric" chromosome is used)}
}
\value{
List which contains results of NSGAII:

\code{values} - Matrix with objective functions values for nondominated solutions.
Each row represents one nondominated solution and each column one objective function.

\code{nondominated_solutions} - Chromosomes of nondominated solutions

\code{statistics} - Statistics about run of genetic algorithm

\code{parameters} - Parameters of genetic algorithm
}
\description{
Use NSGAII algorithm to solve the multiobjective optimization problem.
}
