% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-dominance.R
\name{pareto_dominates}
\alias{pareto_dominates}
\title{Check pareto dominance}
\usage{
pareto_dominates(a, b, minimize = TRUE)
}
\arguments{
\item{a}{Numeric vector with values of objective functions for solution \code{a}}

\item{b}{Numeric vector with values of objective functions for solution \code{b}}

\item{minimize}{If objective functions should be minimized}
}
\value{
Logical value
}
\description{
Check whether solution \code{a} dominates solution \code{b}
}
\examples{
solutionA <- c(1,2,2)
solutionB <- c(5,2,2)
pareto_dominates(solutionA, solutionB)

}
