% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_objective_benchmarks.R
\name{rastrigin_function}
\alias{rastrigin_function}
\title{Rastrigin benchmark function}
\usage{
rastrigin_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Function value
}
\description{
Rastrigin benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
