% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_chromosome.R
\name{scale_numeric_chromosome}
\alias{scale_numeric_chromosome}
\title{Scale values of numeric chromosome to the specified range}
\usage{
scale_numeric_chromosome(chromosome, min = 0, max = 1)
}
\arguments{
\item{chromosome}{Numeric chromosome}

\item{min}{Minimal value(s) which item in chromosome can represent}

\item{max}{Maximal value(s) which item in chromosome can represent}
}
\value{
Numeric vector with converted chromosme values
}
\description{
Scale values of numeric chromosome to the specified range. Numeric chromosome
is a numeric vector with values within [0, 1].
}
\examples{
# Generate numeric chromosome with 10 items
chromosome <- runif(10)
# Scale values of numeric chromosome to [-5, 5]
scale_numeric_chromosome(chromosome, -5, 5)

}
