% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{schaffer_function}
\alias{schaffer_function}
\title{Schaffer function 1.}
\usage{
schaffer_function(x)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Vector with two objective functions values
}
\description{
Schaffer benchmark function 1. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
