% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_objective_benchmarks.R
\name{sphere_function}
\alias{sphere_function}
\title{Sphere benchmark function}
\usage{
sphere_function(v)
}
\arguments{
\item{v}{Vector with input parameters}
}
\value{
Function value
}
\description{
Sphere benchmark function. For more info see
\url{https://en.wikipedia.org/wiki/Test_functions_for_optimization}.
}
