% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigted_sum_ga.R
\name{weighted_sum_ga}
\alias{weighted_sum_ga}
\title{Weighted sum genetic algorithm}
\usage{
weighted_sum_ga(objective_functions_list, weights, chromosome_size,
  chromosome_type = "binary", population_size = 100,
  number_of_iterations = 100, elitism = TRUE, nc = 2,
  mutation_probability = 0.05, uniform_mutation_sd = 0.01)
}
\arguments{
\item{objective_functions_list}{List of objective functions}

\item{weights}{Objective functions weights}

\item{chromosome_size}{Size of chromosome which represents candidate solutions}

\item{chromosome_type}{Chromosome type ("binary" or "numeric")}

\item{population_size}{Number of solutions evaluated in one iteration of genetic algorithm}

\item{number_of_iterations}{Number of iterations (generations) of genetic algorithm}

\item{elitism}{Use elitism}

\item{nc}{NC for SBX crossover (valid if "numeric" chromosome is used)}

\item{mutation_probability}{Probability of mutation (valid if "binary" chromosome is used)}

\item{uniform_mutation_sd}{Standard deviation of mutation (valid if "numeric" chromosome is used)}
}
\value{
List which contains results of weighted sum genetic algorithm:

\code{value} - Sum of weighted objective funtions values for the best solution

\code{best_solution} - Chromosome which represents the best solution

\code{best_solution_index} - Index of the best solution in population

\code{statistics} - Statistics about run of genetic algorithm

\code{parameters} - Parameters of genetic algorithm

\code{values} - Values of objective functions for the best solution

\code{weighted_values} - Values of objective functions multiplied by
weights for the best solution
}
\description{
Use weighted sum approach to solve multiobjective optimization problem.
Weighted sum approach transforms multiobjective optimization problem
to single objective by multiplying objective functions values by weights.
Then genetic algorithm is used to find optimal solution.
}
